// src/app/types/survy.tsx
export interface Option {
    text: string;
    votes: number;
  }
  
  export interface Question {
    id: number;
    questionText: string;
    options: Option[];
  }
  
  export interface Survey {
    id: number;
    title: string;
    questions: Question[];
    status: 'open' | 'closed'; // חייב להיות אחד מהערכים הללו
  }
  